Add support for venvs
If you're in a venv, it shows the name in the prompt. Once I saw a cool
plugin that shows the python version, but that's got a heavy performance
overhead and I'd get annoyed looking at the same number every day.

For the new style to take effect, it expects you to have disabled your
normal venv PS1 by flag (`VIRTUAL_ENV_DISABLE_PROMPT`).